+++ lib/tracking-client.rb
@@ -13,34 +13,49 @@ module Tracking
     SUPPORTED_RESOLUTIONS = %w{year month week day hour}
     TRACKING_KEY_PREFIX = "qs_tracking_"
 
-    def initialize(redis_url)
+    def initialize(redis_url, options = {})
       @minuteman = Minuteman.new(redis: {url: redis_url, driver: :hiredis}, silent: suppress_errors?, time_spans: SUPPORTED_RESOLUTIONS)
+      @options = options
     end
 
-    def track_unique(events, uuid, time = Time.now)
-      Array(events).map do |event|
-        Thread.new {@minuteman.track(event.to_s, UUIDHelpers.uuid_to_int(uuid), time)}
-      end.each(&:join)
+    def track_unique(events, uuids, options = {})
+      time = options[:time] || Time.now
+      uuids = Array(uuids)
+      action = Proc.new {
+        Array(events).map do |event|
+          @minuteman.track(event.to_s, uuids.map {|uuid| UUIDHelpers.uuid_to_int(uuid)}, time)
+        end
+      }
+
+      run_action(action, options)
     end
 
-    def query_unique(event, resolution, time = Time.now)
+    def query_unique(event, resolution, options = {})
+      time = options[:time] || Time.now
       raise Error.new("Resolution `#{resolution}` not supported!") unless resolution_supported?(resolution)
 
       result = @minuteman.send(resolution, event.to_s, time)
       QueryResult.new(result)
     end
 
-    def track_impression(events, time = Time.now)
-      Array(events).map do |event|
-        Thread.new do
-          Impression.new(event, time).keys.map do |key|
-            Thread.new {redis.incr key}
-          end.each(&:join)
-        end
-      end.each(&:join)
+    def track_impression(events, options = {})
+      time = options[:time] || Time.now
+
+      action = Proc.new do
+        Array(events).map do |event|
+          Thread.new do
+            Impression.new(event, time).keys.map do |key|
+              Thread.new {redis.incr key}
+            end.each(&:join)
+          end
+        end.each(&:join)
+      end
+
+      run_action(action, options)
     end
 
-    def query_impression(event, resolution, time = Time.now)
+    def query_impression(event, resolution, options = {})
+      time = options[:time] || Time.now
       redis.get(Impression.new(event, time).key_for(resolution)).to_i
     end
 
@@ -55,6 +70,18 @@ module Tracking
 
     def redis
       @minuteman.redis
+    end
+
+    def run_action(action, options)
+      if synchronous?(options)
+        action.call
+      else
+        Thread.new(action) {|action| action.call}
+      end
+    end
+
+    def synchronous?(options)
+      options.has_key?(:blocking) ? options[:blocking] : @options[:blocking]
     end
   end
 end